requirements.txt
asgiref==3.8.1
astroid==3.3.9
dill==0.4.0
Django==5.2
django-appconf==1.1.0
django-ckeditor==6.7.2
django-compressor==4.5.1
django-js-asset==3.1.2
django-redis==5.4.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
isort==6.0.1
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.1.0
rcssmin==1.1.2
redis==6.0.0
rjsmin==1.2.2
sqlparse==0.5.3
tomlkit==0.13.2
types-PyYAML==6.0.12.20250402
typing_extensions==4.13.2
whitenoise==6.9.0
